requirements.txt
streamlit
fastapi
langchain-google-genai
python-dotenv
langchain
langchain-community
beautifulsoup4
qdrant_client
langchain-qdrant
Scrapy
regex
unstructured
selenium
playwright
python-docx
python-multipart
uvicorn
langchain-experimental
scikit-learn
firebase-admin
docx2txt
pylatex
PyPDF2
gtts
boto3
pydub
sympy
pyttsx3
pillow
python-multipart
oauth2client
google-auth-oauthlib
openpyxl
xlrd
tabulate